requirements.txt
Django==4.2.11
djangorestframework==3.15.2
drf-yasg==1.21.7
python-telegram-bot==20.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9 
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
python-telegram-bot==20.7